Fix casting for Windows. (strlen)
git-svn-id: file:///home/jj/hercules.svn/trunk@6466 956126f8-22a0-4046-8f4a-272fa8102e63

+++ cgibin.c
@@ -1208,7 +1208,7 @@ void cgibin_cmd_cmd(WEBBLK *webblk, char *command)
     html_header(webblk);
     hprintf(webblk->sock, "<PRE>\n");
 
-    cgibin_hwrite(webblk, response, strlen (response));
+    cgibin_hwrite(webblk, response, (int)strlen (response));
 
     hprintf(webblk->sock, "</PRE>\n");
     html_footer(webblk);
